flask-server/requirements.txt
Flask==2.1.1
Flask_PyMongo==2.3.0
pymongo==4.1.1
python-dotenv==0.20.0
spacy==3.0.6
